refactor code
- move constants to separate files
- move header and related files to separate directory
