+++ src/components/About/Header/MobileNav.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import { LINKS, SOCIALS } from '../constants'
+
+const MobileNav = ({ setShowNav }) => {
+  return (
+    <div className='bg-white fixed flex flex-col h-full items-center justify-center left-0 top-0 w-full'>
+      <button
+        className='block fixed m-4 right-0 rounded-full px-3 py-2 text-2xl top-0
+          focus:bg-gray-500 focus:bg-opacity-50 md:hidden'
+        onClick={() => setShowNav(false)}
+      >
+        <i className='ri-close-line'></i>
+      </button>
+      {LINKS.map((link, index) => (
+        <a
+          key={index}
+          className={index === 0 ? 'm-3 text-green-500 text-xl' : 'm-3 text-xl'}
+          href={`#${link}`}
+        >
+          {link}
+        </a>
+      ))}
+      <div className='mt-3 text-xl'>
+        {SOCIALS.map((social, index) => (
+          <a key={index} href={social.link} rel='noreferrer' target='_blank'>
+            <i className={`mx-3 ${social.icon}`}></i>
+          </a>
+        ))}
+      </div>
+    </div>
+  )
+}
+
+export default MobileNav
