+++ src/components/About/constants.js
@@ -1,3 +1,5 @@
+export const LINKS = ['about', 'courses', 'projects', 'blog']
+
 export const PROFESSIONAL_EXPERIENCE = [
   {
     duration: '< 1 year',
@@ -10,5 +12,24 @@ export const PROFESSIONAL_EXPERIENCE = [
     logo: 'https://upload.wikimedia.org/wikipedia/commons/thumb/3/34/EY_logo_2019.svg/1200px-EY_logo_2019.svg.png',
     subtitle: 'Pen Testing, Burp Suite, Qualys',
     title: 'Cybersecurity Analyst',
+  },
+]
+
+export const SOCIALS = [
+  {
+    icon: 'ri-discord-line',
+    link: 'https://discord.gg/PA5ZxAuSp3',
+  },
+  {
+    icon: 'ri-linkedin-box-line',
+    link: 'https://www.linkedin.com/in/jagannathbhatjb/',
+  },
+  {
+    icon: 'ri-github-line',
+    link: 'https://github.com/jagannathBhat',
+  },
+  {
+    icon: 'ri-gitlab-line',
+    link: 'https://gitlab.com/jagannathBhat',
   },
 ]
